test: refactor parametrized tests
- The parser tests are split in two modules: one for domain's, the other for problem's parser tests.
- The parametrization over PDDL fixture files of the parser tests is changed. Now we iterate over all problems, and for each problem we also parse the respective domain, and then we check the validity of the problem against the domain (via the implicit call to the Problem.check method, triggered by the Problem.domain property setter).
